requirements.txt
python-telegram-bot>=21.0
anthropic>=0.40.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
yt-dlp>=2024.0
rumps>=0.4.0
